feat(techaer C of CRUD): creation of teacher and classes

+++ src/useCases/school/deleteSchoolUseCase.ts
@@ -6,6 +6,14 @@ interface DeleteSchoolUseCaseRequest {
 
 export class DeleteSchoolUseCase {
   async execute({ id }: DeleteSchoolUseCaseRequest) {
+    if (!id) return;
+
+    await prisma.classroom.deleteMany({
+      where: {
+        schoolId: id,
+      },
+    });
+
     await prisma.school.delete({
       where: { id },
     });
